move after_log as general one
